Refactor time_data to use Null Object (#996)
I had to un-memoize time_data largely to deal with factories and specs where @time_data was getting called in the set_problem_order callback before the reservation was added. It should be quick enough that it won't really matter.

One thing this does is remove the delegation from Reservation#problem? to OrderDetail, which is only used in specs, and aliases requires_but_missing_actuals? to problem?. Occupancies will need to implement problem?.