[#144318485] Refactor instrument price policy calculations (#997)
Refactors the instrument price policy calculations so a good chunk of the logic is now in a separate class which I plan use for the secure rooms price policies calculations.

There's still more refactoring that could be done (I hate the interface for price policy calculations, but I didn't want to break too much). I'm not sure how much more I'll do on that side right now, but there's definitely specs to be moved or written.