Foreign keys and generating ERD diagrams
This PR 

* Adds the rails-erd gem which allows us to create database relationship diagrams

* Adds a bunch of foreign keys and indexes

* Fixes some tests that break with foreign keys

My research of all the things that look like foreign keys but don’t have foreign keys defined is at:

https://docs.google.com/a/tablexi.com/spreadsheets/d/1fbVPc9qr1iqirQLvJNTSflpyOCb-Fuitr-4FetE3ZEU/edit?usp=sharing

In creating foreign keys, I did not use all the possible ones, I tried to pick the ones that were in join tables, or involved users, or seemed to be significant.

I attempted to run the migration against recent production of the UIC and Dartmouth databases, the migration file is commented with the lines that cause problems in those databases.