[#144058951] Complete orders for entry only rooms (#1051)
If there is an entry-only room (no exit scanner), then go ahead and complete the order as is.

Once the occupancy is created, it can be given a duration and exit at. I tried to see about making the form only have the single time for entry-only rooms, but dealing with updating the exit time to match the entry time whenever you change it proved to be beyond of what I consider the expected effort.