[#144058951] Use Event#occurred for occupancy times
When associating both an entry_at and exit_at for an entry-only
scanner, the timing of the two `associate_` calls mattered because if
`associate_exit` happens first, then `entry_at > exit_at`, which causes
the price policy to fail. This instead uses the event’s `occured_at`
time to set both, so the times will be equal.