[#136429] Prevent duplicate details in Journal#order_details (#1079)
When there are orders with a SplitAccount on a journal, the association was coming back with duplicate order details because there are multiple journal_rows per order detail (multiple journal rows is intentional). This works fine on MySQL because of the `distinct` scope.

However, the Oracle-specific `clob_safe_distinct` was not working because the inner `distinct` was not being scoped, so it was essentially coming out as `WHERE id IN (SELECT DISTINCT(*) FROM order_details)`. Using a normal `distinct` has problems in oracle because `DISTINCT *` does not work with CLOB fields (notes). `ORA-00932: inconsistent datatypes: expected - got CLOB 00932. 00000 -  "inconsistent datatypes: expected %s got %s"`.

I've made the existing association essentially private with a note that it may contain duplicates. There is now a regular method acting as a pseudo association for `Journal#order_details`