Skip adding cancelation fields to reservations with missing order details
There are about 46 order_details in UIC that were deleted, but the reservations were
not. I suspect this was back when they were running MySQL 5.5, and possibly on MyISAM,
which does not support foreign keys. Likely, when they exported from the old
server and imported to the newer one, the importer temporarily disabled foreign
keys.