[#136705] Disputed orders not displaying in Notices (#1125)
for facility admins. In the upper right "Notices" dropdown, "Disputed Orders" was no longer appearing.

When we got rid of the separate Disputed tabs on "Orders" and "Reservations". When we did that, we got rid of the permission those tabs were using. https://github.com/tablexi/nucore-open/pull/1059

The first commit is the simple change, but it was clear the way we were testing might have been a little too close to the code itself, which was how we missed the regression. So I started to refactor the specs to use actual users with a role instead of a stubbed ability. This led me to also refactor a little bit of the code itself. I also found were one of the specs was actually incorrect (it wasn't a big deal because it involved the training request feature being disabled, but there being a training request in the db).

For context, see https://github.com/tablexi/nucore-open/commit/bae6e5388ef893c01fcadf1d1bc2e79968225e66 for how this is expected to behave.