[#136702] Prevent invalid quantities from cart into Sanger form (#1123)
In the cart, if the user changes the quantity field to be blank or negative and then clicks the "Complete Order Form" we end up passing a blank string or negative number in the quantity query parameter to the Sanger form. This was then triggering an `ArgumentError: quantity must be positive` in `Submission#create_samples!`.

I also tested with decimal numbers. Those were being passed as decimals, but the form was handling that by dropping the fractional part.

This now handles it at both the cart level and the create_samples level. The cart will disable the "Complete Order Form" button if the quantity is not a valid value. If an invalid value does slip through, then the Sanger form will default to having one sample.