[#136759] Use ReservationUserActionPresenter for OrderDetails#show
The “Cancel” button had been broken since the beginning of git history:
the “Initial Fork”. This uses the same presenter the “My Reservations”
page uses, so we’ll get all the same options.

This also pulls the joining of the actions out into the helper. I
originally did it because I was going to use multi-lines or `ul` in
this view, but decided to stick with the pipe-delimited. I left the
change in because I do think it’s more intuitive if the presenter spits
out an array of links.