[#136792] Stricter date format from user input
Force to always use 4 digit years. Before, “3/4/17” was being treated
as  the date “0003-4-17”.

Since the method returns nil for invalid dates, this could still cause
unexpected behavior. For example, you can backdate an order, but if the
input is blank or invalid, it defaults to Time.current. This could
result in setting the ordered_at to now rather than what the user
expected to see.