[#136792] Consider 2-digit years as an invalid format (#1134)
* Stricter date format from user input

Force to always use 4 digit years. Before, “3/4/17” was being treated
as  the date “0003-4-17”.

Since the method returns nil for invalid dates, this could still cause
unexpected behavior. For example, you can backdate an order, but if the
input is blank or invalid, it defaults to Time.current. This could
result in setting the ordered_at to now rather than what the user
expected to see.

* [#136792] Validate date format in date picker widgets

This highlights the form element in red if the date format is invalid
(only happens if the user manually adjusts the date w/o using the
widget).

I considered doing something like disabling the form, but that started
to get weird with also wanting to disable the submit buttons when
thinking about multiple things possibly enabling/disabling.

I’m hoping just highlighting the problem will dissuade users from
hitting submit.