[#136869] Unable to add auto-scaled accessory to running reservation (#1143)
* Create & Start a reservation on an instrument with accessories
* While it is running, click "Add Accessories"
* Manually scaled accessories are defaulting to "0:00". It should
default to the current duration
* Auto scaled accessories show "0:00", and then when you try to save,
it says "must be greater than or equal to 1"

Adding accessories works fine once the reservation has ended. That's because these are querying the reservation for its `actual_duration_mins`, which now is `nil` if the reservation is running or a problem. This adds a new method that allows you to fall back to using the current time like older behavior. See #1047.

![screen shot 2017-06-26 at 6 54 48 pm](https://user-images.githubusercontent.com/1099111/27614371-cea520a2-5b66-11e7-9846-7ba7e3e04a6f.png)