[#136900] User details being shown on calendar
There is a `show_details` “Show reservation details (name, email, etc)
on schedule” option on instruments that should only show users’ name
and email if the option is on.

However, what was happening is we were sending the string `”false”`
into `Reservations::CalendarPresenter`, which is `present?`, so we were
rending the details if if we didn’t want to.

Even if that was fixed, using a query parameter like this isn’t very
secure. Anyone could conceivably change the query parameter to “true”.
This changes the logic so that if you request details (which happens in
most of the facility-side views), it also checks to make sure that
either the `show_details` option is on or that you have facility-level
permissions.

I also did a little bit of JS cleanup getting rid of the `withDetails`
javascript global since it’s only used in the one case of a user
placing a reservation to disable the tooltips.