[#136306] Fix reservation validations (#1153)
Applies to NU's [#136043] for research safety and UIC's required note field.

If you have multiple line items in your cart that have validations including a reservation, when you try to create the reservation, it is validating every other line item in your cart.

So, add an item that requires a note and an instrument that doesn't to your cart. Click the "Make a Reservation" button. You will not be able to save the reservation, and you will get a "Validation failed:" flash. Part of this is because for reservations, we're displaying the order details's validations.

This changes the note and research safety validators to accept a list of order details (or single order detail) rather than an order. It also changes the research safety validator to place errors on the model itself.

These commits have been cherry-picked from tablexi/nucore-nu#283