Fix Date to Time conversion so it uses the app's timezone (#1155)
`to_time` uses the system's timezone, not the app's.

This was broken by changes made in https://github.com/tablexi/nucore-open/pull/1134

Circle is set to central time as is the application, so that's why this was fine until we merged to DC, which has circle set to central and the app set to eastern. I'm going to open another PR for trying to deal with that.