[#137451] Extend journal row description field length (#1204)
`OrderDetail#long_description` is what gets put into the
`JournalRow#description` field. If you have a really long product name
(that field is limited to 200 characters), then the description is too
long for the journal row field.

The 512 is somewhat arbitrary, but should be more than enough for
fitting long product names, the order detail and a user’s name.