[#137409] Handle invalid dates for reservations on completed
Looks like we had fixed the non-completed scenario in #1201, where it
goes through estimate_cost_and_subsidy, but not the completed order
scenario where it goes through calculate_cost_and_subsidy.

I think we didn’t have the guard clause here because it was assumed
that we would only ever try to calculate pricing on a valid order
(which would always have these fields), but in the `PriceChecker` used
on the order detail popup as an ajax call, it’s possible for the
reservation to become invalid if there is an error parsing the date.