[#137105] Persist note on updating a reservation (#1218)
We did the front end piece in
https://github.com/tablexi/nucore-open/pull/1208 but didn’t verify it
actually persisted.

This also adds the note required validation so the user can’t remove
it. It would be possible to abstract the validation out, but we
probably want this list to be different from the on-purchase
validation. It just didn’t seem worth it right now. If we start adding
other validations (for example, I don’t think we need to run the NU
research safety validations), then it might be at that point.

This largely works because because Reservation delegates `note` and
`note=` to the order detail so in this case they can be treated
interchangeably.