[#139318] Ignore ActionController::UnknownFormat in ExceptionNotifier (#1245)
There's no real need to be notified of this error. This happens if a js/json path meant for ajax is called without a format header. It is already ignored by Rollbar.

This also ignores `ActionController::UnknownHttpMethod` in rollbar. We sometimes see those errors when we're getting vulnerability scanned.

This will get a little funky going into the schools. NU already has the `ActionController::UnknownHttpMethod` ignore and Dartmouth doesn't use ExceptionNotifier. But this should help get things closer in line.