[#136545] Add Filtering to Reconciliation Pages (#1294)
As part of this feature, this creates a parallel path for doing transaction searching. The end goal is this will replace the current `TransactionSearch` module that is included into controllers and used with the `_with_search` action. That will give us more control over each individual action while also hopefully much easier to reason about.

That is why there is a `_search2` view and a `TransactionChosen2Input`. Ideally, they will replace the original, but it seemed more prudent to create this parallel method than to try to refactor the original.

![screen shot 2017-12-20 at 12 40 28 pm](https://user-images.githubusercontent.com/1099111/34221546-62da1b0c-e586-11e7-9e23-fddb3ed9729a.png)

The new chosen input has some additional features that will come to play later like being able to render `data` attributes on individual `options`. This is used in the product filter. I did not leave the product filter in this PR, because we're not using it here, but I did test with it.