[#139830] Fix schedule rule discount precision
Here's an example of what was happening from a local test:

* $60/hour ($1/minute) price policy
* 24 hour/day schedule rule with a 10% discount.
* Time without relay
* Started 12:34:09, ended 12:41:41

This should be calculated for 7 minutes, so $7 with a 10% discount = $6.30. Instead it is coming out as $6.26.

The way we calculate a discount is to find the effective discount percent. We find the percent of time a reservation overlaps with each schedule rule, multiply by the discount and sum them up. In this case we have the one rule which should be 100% * 10%, but because we're not trimming the seconds in this part of the calculation we're using 7.53 minutes, it's coming out to 106% overlap with the schedule rule. We then multiply that by the discount (10%), so we come out with as a 10.619% discount.

I know that logic is confusing, so it might make more sense to see a normal case when you span multiple schedule rules with different discounts. An example:
* 9am-5pm schedule rule with a 10%
* 5pm-midnight 20% discount
* Reservation runs from 4:30-5:30.

30 minutes is in rule 1 (50%)
30 minutes in in rule 2 (50%)
50% * 10% + 50% * 20% = 15% effective discount