[#139829] SecureRooms: Always prompt for account selection, even if only one
So that users always know exactly which account is going to be charged, we
want to always display an account selection, even if there is only one account.

With this change, there is no need to change anything on the nucore-rooms elixir
app or mobile app: as long as we respond with "pending" and the account list (of
only one thing), they behave appropriately.