Replace "All Transactions" search with newer search method (#1295)
This continues from #1294 and replaces the "All Transactions" search logic, e.g. http://localhost:3000/facilities/example/transactions with the new search. It gets rid of the `_with_search` metaprogramming. Hopefully, this is easier to reason about.

The feature spec I wrote works for both master and this branch. I think more specific unit tests are going to be better for testing the search criteria than the existing controller specs, which is why I did one happy-path feature spec and killed the controller specs for this action.