[#140459] Fix display when no transactions on All Transactions
Broken by #1390, if there were no transactions matching the default
search on All Transactions (fulfilled since the beginning of last
month), the search form was not rendering.

Now, we will always render the form. This means the form might have no
values in any of the selects, but that should only happen until the
facility gets its first order.