[#137552] Better error message when beginning a reservation in grace period
When there is a currently running reservation, improve the error
message when you try to “Begin Reservation” in the 5 minute grace
period. Once the grace period is over (i.e. the previous reservation is
technically over), “Begin Reservation” will start the reservation and
move the prior reservation to the problem queue