[#141665] Support both unix timestamps and iso8601 in calendar endpoint
For some reason, we’re continuing to get `reservations#index
(ArgumentError) "argument out of range”` errors caused by start/stop
parameters being in unix timestamp format. This format changed when we
upgraded Fullcalendar, so it seems like some browsers are continuing to
load the old code. That shouldn’t be happening because the library
should be fingerprinted.

This should stop the errors.