Allow admin-defined logo in header
This change allows administrators to add a logo to their header. The new
logo image should be called "logo-header.png" and it should be placed in
the "app/assets/images/" folder. The changes are not visible
immediately, only after the asset pipeline is rebuilt (which just
involves restarting Rails).

The diff is a little interesting because the header was redone in
Flexbox. The new version is pixel-for-pixel the same, though, and it
shouldn't affect user overrides as far as I can tell.