Fix oracle types in Rails 5
In rails 4.2, ruby `DateTime` maps to an oracle `DATE` column (which supports seconds, but not subseconds). Now it's assocated to ruby `Date`. So In rails 5, it's only grabbing the date portion of those columns. And the schema dumper writes them as  `t.date` rather than `t.datetime` https://github.com/rsim/oracle-enhanced/pull/875

Looks like they're going to be using TIMESTAMP for the rails datetime. https://github.com/rsim/oracle-enhanced/pull/739

This ends up being a no-op when applied to MySQL.

For additional reference, we go from 1.6.9 -> 1.7.11 of the oracle adapter https://github.com/rsim/oracle-enhanced/blob/master/History.md

I kind of like changing the columns in that it gives us fractional sections, which isn't important on its own, but does give closer parity to mysql.

Also, I ran this against a production backup. It was super fast; I was worried this would be slow on the really big tables.