Update LDAP documentation (#1729)
# Release Notes

Update documentation for developing against LDAP

# Additional Context

I did this while starting to work on the tool for auto-suspending users if they are removed from the LDAP directory.

Removes older instructions for setting up an LDAP server on your Mac OS system. We now recommend using Docker.

This also adds the ability to use YAML aliases in the configuration.