[#143771] Prevent error on invalid dates on Reservation#update (#1731)
# Release Notes

Prevents an error on the reservation edit form when start/end at is blank because of invalid data.

# Additional Context

Here's the error: "NoMethodError: undefined method `<' for nil:NilClass" on the `if duration_is < duration_was` line. 

The error that we originally got notified of was super weird: it was missing the `reserve_start_mins` and `reserve_end_mins` parameters. I could recreate that if I deleted the dropdown from within my browser's dev tools. But I was able to replicate it as well by putting in an invalid date like "10/0/2018" into the date fields.