[#140549] Retry LDAP queries up to 3 times (#1770)
# Release Notes

Tech task: Retry, up to three times, LDAP queries that time out.

# Additional Context

This should help with the LDAP deactivation job getting occasional timeouts which
was causing the entire job to fail. It'll still end up crashing if there's a serious issue with LDAP, but hopefully, this will handle most of the cases.