[#145520] Restore ability to have links open in new window (#1781)
# Release Notes

Restore ability to allow links in product and facility descriptions to open in a new window.

# Additional Context

Users have the ability to set a `target` on their links in the facility and product descriptions through the editor to allow opening a link in a new window. We added some sanitization to `Facility#description` in #1703 which strips out that attribute. This lets it through sanitization.