Handle External Service URLs that already contain a question mark (#1789)
# Release Notes

Handle external survey links that already contain a question mark.

# Additional Context

If a survey link has a question mark in it, we would append additional query params with our own `?`. This came up in the context of PerfectForms. Qualtrics and Surveyor don't have `?` in their URLs, so it wasn't a problem before.

`http://www.survey.com/surveys/edit?id=123` would end up as `http://www.survey.com/surveys/edit?id=123?order_id=123-456&...`. This corrects it so we only end up with a single question mark.

This was originally done in #1771 and reverted in #1780 because some of the URLs on NU prod contained leading whitespace which caused `URI(url)` to raise a `ActionView::Template::Error: bad URI(is not URI?)` error.