[#145115] Add minimum price change reason length for order details (#1797)
# Release Notes

Require that the price change reason for an order be at least 10 characters. This validation only applies when an order is being manually adjusted from its calculated cost.

# Additional Context

I debated if it was worth the effort of moving the specs over to the model, but they were already in the controller spec and the model spec is already almost 2000 lines long.

Also, by moving the `feature_setting` tag, I was able to drop the time that file takes to run from about 2:40 to 1:30 (flipping the feature settings is slow because we also end up reloading the routes).

UIC is the only school with this feature disabled, so we might be able to remove the feature flag at some point in the future.