[#144522] Fix bundle column on export raw (#1811)
# Release Notes

Fix bundle column in export raw to properly show the bundle name if the order was purchased as part of a bundle. The column was previously called "Bundled Product".

# Additional Context

Currently, if you place an order with a bundle and then go to find those line items on the export raw report the "Bundled Products" column is always blank. I'm guessing that at one point (way back to even before my time) that a bundle was a single line item with multiple products. That appears to be what it's trying to report on. For a long time now, when you add a bundle, each product of the bundle is a separate line item with a `bundle_id` tying back to the product.