Use-Manage Facility List (#1803)
# Release Notes

Add facility thumbnail upload capability, as well as a tiled facility list. This works with css grid and has a flexbox fallback.

# Screenshot
## Rendered tiles
<img width="1213" alt="screen shot 2018-12-13 at 12 33 54 pm" src="https://user-images.githubusercontent.com/5000430/49956646-aedbe000-fed3-11e8-953c-a38c769aa17b.png">
As visible, EF has a thumbnail while FAC2 does not.


## Upload field well
<img width="910" alt="screen shot 2018-12-13 at 12 45 06 pm" src="https://user-images.githubusercontent.com/5000430/49957152-f151ec80-fed4-11e8-8cba-320f37b7eeb4.png">
Uploading a thumbnail is done in the admin form.

# Additional Context

This includes use of a feature flag and works with the az-list. I needed imagemagic installed on my machine for this to work.