Tech task: Fix sporadic test failure (#1822)
# Release Notes

Tech task: Fix sporadic test failure

# Additional Context

The users in the controller are sorted by last name, first name. The user factory
assigns last names of integers (e.g. "8", "9", "10"), so if it happens to hit where
it goes from 9 to 10, the ordering may not be what the test was expecting. We don't
really care about the ordering in this spec, so we can ignore it.

See https://circleci.com/gh/tablexi/nucore-open/7099