[#146242] Fix for split accounts when cancelled with a fee (#1826)
# Release Notes

Fix errors on Create Journal and report when there is a canceled with cost reservation on a split account.

# Additional Context

This is a side effect of the new calculated cost fields in #1810. All of the other cost calculations return an integer or BigDecimal, but cancelation fee was returning a Float, which can't be converted to a BigDecimal w/o a precision (`can't omit precision for a Float`).