Only split calculated_cost when doing export raw (#1833)
# Release Notes

Only calculate `calculated_cost`/`subsidy` when splitting an order detail if we are in the context of export raw.

# Additional Context

There are a few other places like the the Journals#new (where we partition into valid/invalid accounts) that split order details. In these scenarios, the calculated_cost isn't used, so we don't need to copy it. This should add a performance improvement as well as avoiding errors like #1826.

Those columns were added in #1810.