[#144419] Interface for managing an Account's facilities (#1840)
# Release Notes

Add an interface for global admins to manage which facilities an account is valid for. This applies to non-global accounts (e.g. Credit Cards and POs).

# Screenshot

![screen shot 2019-01-16 at 4 14 13 pm](https://user-images.githubusercontent.com/1099111/51282192-12ad6900-19aa-11e9-808e-df8b69bc11ff.png)

# Additional context

I considered using checkboxes, but this seemed better for when there are lots of facilities.

We will probably want to disable this for NU for a little while until they can get BI to update their queries.