Always return Float from Reports::InstrumentUtilizationReport::DataRow#safe_divide
In Ruby 2.5, `Integer#round`’s behavior changed to return `self` when that method’s argument is positive. As a result, if we call `safe_divide(a, b)` when `b` is zero, and we then call `.round(1)` on that result, we now get 0 instead of the expected 0.0.

To get predictable behavior, we always return a Float from `safe_divide`.