Show all facilities for account (#1868)
# Release Notes
Fix display of multi-facility accounts in the Manage Facilities > Users > User > Payment Sources tab.

# Additional Context
I thought I had done this already, but it might have gotten lost in one of my branches. I'm surprised this doesn't error on #1863. I may try to add a spec over there to catch this.

This could get unwieldy if they start adding too many facilities, but I think we can address it if it becomes a problem. I was thinking maybe truncating after a certain number, so "Facility 1, Facility 2, Facility 3, and X more..." or something like that.