Only act in Reservation#set_billable_minutes if a price policy is present
In several places, we save a reservation before updating its order detail, while we are making the order detail complete (at which point we assign it a price policy). Hence, it is unfortunately required to check for the presence of a price policy, even though an order detail should never be able to be complete and without a price policy at any instant in time.