[#152015] Fix error on receipts containing sanger results files
`ActionView::Template::Error: undefined method `facility_sample_result_path'` because the path helper is not available in the mailer, but there was a partial that was being shared between regular web views and the email.

This should normally not be an issue since the receipt is usually sent immediately on purchase, so logically it hasn't been fulfilled and no results files have been uploaded. However, on the facility admin's order view, there is a "Send Receipt" button which re-triggers the email, so there is a possibility of this error being triggered in the future.

It also happened when delayed_job had crashed for a couple days and receipts for fulfilled orders were still queued up.

See https://rollbar.com/tablexi/nucore-dartmouth/items/138/ and https://rollbar.com/tablexi/nucore-dartmouth/items/162/