[#152079] Fix adding timed service bundle (#2141)
# Release Notes

Fix an issue where a timed service as part of a bundle cannot be added to an existing order. It was getting stuck as if there were an issue like a missing order form.

# Screenshot

What was happening before:

![All Bundle Add to Order](https://user-images.githubusercontent.com/1099111/71695338-69f14880-2d77-11ea-8c0c-8b79c602abd5.png)

# Additional Context

We fixed a hard error in #2128, but this was still having a problem while validating the order.