Research safety refactors for UMass (#2134)
# Release Notes

Tech task: refactor parts of research safety to make it better for UMass.

# Additional Context

The big thing this does is allow the forks to do their own internal caching of API calls. UMass will be doing a single call per user which returns a list of all of the active certifications. With this change, it uses the same object for multiple `certified?` checks so the HTTP response can be memoized.

I refactored this in conjunction with https://github.com/tablexi/nucore-umass/commit/fa658ac6cfc784f2a2d4be1298393ea6a1d52122 (which is not ready due to us still needing some additional data from them).

It also sets up a default that assumes a user is always certified. We'll need to swap this out in NU and UMass.