Fix recharge account editing
Editing a recharge account was broken by a recent change and turns out we didn't have a good spec around it.

The spec is a little funky in how it fills out the form, but it needs to be in order for NU and Dartmouth to pass since they use different sets of fields.